feat: Use flex gap for Cluster
BREAKING CHANGE: This is a breaking change, as it removes the outer wrapper and thus changes the shape of the HTML that is emitted.

It is a much cleaner/nicer solution though, and supported in all modern browers (Safari 14.1+).

There is a fallback behavior that is detected via JS (as it cannot be properly detected in CSS alone). The fallback will ensure there are still gaps, but importantly will _not_ use the negative-margin hack to ensure the items are flush on the outside. 

The fallback behavior should be acceptable, not perfect but workable.